+++ database/migrations/2020_12_23_233134_create_player_timing_stage_table.php
@@ -22,7 +22,7 @@ class CreatePlayerTimingStageTable extends Migration
             $table->double('TimeInZone');
             $table->unsignedInteger('Attempts');
             $table->double('Sync');
-            $table->double('Speed');
+            $table->unsignedInteger('Speed');
             $table->unsignedInteger('Jumps');
             $table->dateTimeTz('CreatedDate')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->dateTimeTz('LastModifiedDate')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
